Create Board.py
Basic board position class so far. Can add pieces, display its output, and has some really basic legal moves so far. Ideally it will be responsible for moving pieces, looking for moves, and so on. Takes input from the Piece class. The way I have it pictured so far, this class is responsible for quite a bit, so if you want to break it apart into multiple classes feel free.